Update APIs and DB interactions to reflect email confirmation step
